+++ changelog.md
@@ -22,6 +22,7 @@ For example:
 * Added formula round and int functions. [#891](https://gitlab.com/bramw/baserow/-/issues/891)
 * Views can be duplicated. [#962](https://gitlab.com/bramw/baserow/-/issues/962)
 * Link to table field can now link rows in the same table. [#798](https://gitlab.com/bramw/baserow/-/issues/798)
+* Redirect to signup instead of the login page if there are no admin users. [#1035](https://gitlab.com/bramw/baserow/-/issues/1035)
 
 ### Bug Fixes